requirements.txt
numpy==1.20.3
Pillow==8.2.0
rstr==3.0.0
scikit-learn==0.24.2